+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to com.fusedvr.web3 package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2022-04-24
+
+- Re-order appId and email parameters for login
+- Store Token in Player Preferences to be used between play throughs
+- Add Function for Getting Magic Link after Login
+
 ## [0.1.0] - 2022-01-02
 
 - Initial Release of FusedVR Authentication SDK